This should fix #147.
The real explanation comes from the Python 3.2 docs for the subprocess module:

"If close_fds is true, all file descriptors except 0, 1 and 2 will be closed
before the child process is executed. (Unix only). The default varies by
platform: Always true on Unix. On Windows it is true when stdin/stdout/stderr
are None, false otherwise. On Windows, if close_fds is true then no handles
will be inherited by the child process. Note that on Windows, you cannot set
close_fds to true and also redirect the standard handles by setting stdin,
stdout or stderr.

Changed in version 3.2: The default for close_fds was changed from False to
what is described above."

This was changed in Python 3.2 such that the default value for close_fds is no
longer False on Windows.  Although I haven't examined this too closely, I think
that what happens is that since close_fds is now True by default, when we run
the test command in a subprocess, the subprocess does not inherit the actual
stdin/stdout/sderr handles, which is not a scenario expected by py.test's I/O
capture.

Setting close_fds=False is fine on all platforms for our purposes.  I tested
this in Linux too with my staging build.