Some refactoring of astropy.tests and astropy.__init__.  A few things going on here:
1) The TestRunner class has been moved out of astropy.tests.helper
into its own module astropy.tests.runner.

2) The astropy.tests.__init__ module has been cleaned up so that there
isn't anything imported into the top level of the astropy.tests
namespace.  In particular astropy.tests.helper is not imported.  This is
so that astropy.tests.runner can be imported without causing pytest to
be imported too.

3) Added a TestRunner.make_test_runner_in classmethod--this actually
*creates* the test() function that is called as astropy.test().  It
gets its docstring and arguments from TestRunner.run_tests (this is
in contrast to the previous situation where TestRunner.run_tests and
astropy.test got different information from each other, and arguments
signatures had to be specified in two different places).

4) Added astropy.tests.command module containing the AstropyTest
setup.py command implementation (moved back from astropy-helpers).
In retrospect it was a mistake to keep this separately in
astropy-helpers, because the functionality of the test command is very
much tied to what capabilities are available in the test runner, so that
having them in separate projects made it difficult to keep them in sync.
An associated change will be added to astropy-helpers to add a wrapper
that attempts to import the test command from astropy (and will provide
a fallback if it is unavailable).

5) Changed how the astropy.__init__ namespace is cleaned up.