BUG: Fix module level import requires
Since the logic was written in a negate way compared to the normal
skipping, the module level skip didn't take into account the special
"*" and "." patterns.

gh-109 probably misdiagnosed the module actually missing only on the
run the had pytest-xdist: the test does not seem to ever pass either
way.

Closes gh-109